Optional normalised-by-points mode.

+++ src/main/kotlin/io/hsar/mathhammer/model/OffensiveProfile.kt
@@ -3,7 +3,7 @@ package io.hsar.mathhammer.model
 data class OffensiveProfile(
         val name: String,
         val skill: Int,
-        val attacks: Int,
+        val attacks: Double,
         val strength: Int,
         val AP: Int,
         val damage: Int
