+++ src/main/kotlin/io/hsar/mathhammer/statistics/HitCalculator.kt
@@ -1,7 +1,7 @@
 package io.hsar.mathhammer.statistics
 
 object HitCalculator {
-    fun hits(skill: Int, attacks: Int): Double {
+    fun hits(skill: Int, attacks: Double): Double {
         return (skill / 6.0) * attacks
     }
 }
